6-size.c

+++ 0x00-hello_world/6-size.c
@@ -14,11 +14,11 @@ int main(void)
 	long long int d;
 	float e;
 
-	printf("size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
-	printf("size of an int: %lu byte(s)\n", (unsigned long)sizeof(b));
-	printf("size of a long int: %lu bytes(s)\n", (unsigned long)sizeof(c));
-	printf("size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
-	printf("size of a float: %lu byte(s)\n", (unsigned long)sizeof(e));
+	printf("size of a char: %lu byte(s)\n", sizeof(a));
+	printf("size of an int: %lu byte(s)\n", sizeof(b));
+	printf("size of a long int: %lu bytes(s)\n", sizeof(c));
+	printf("size of a long long int: %lu byte(s)\n", sizeof(d));
+	printf("size of a float: %lu byte(s)\n", sizeof(e));
 	return (0);
 
 }
